chore: move styles to a separate folder
Also
- add "waiting for scan" caption for the processing screen
