+++ src/components/web-print/FileDrop.tsx
@@ -1,8 +1,8 @@
-import styles from "@/components/web-print/printers.module.css";
+import styles from "@/components/web-print/css/printers.module.css";
 import { useEffect, useRef, useState, JSX } from "react";
-import fontStyles from "@/components/web-print/printers.fonts.module.css";
-import marginStyles from "@/components/web-print/printers.margins.module.css";
-import themeStyles from "@/components/web-print/printers.theme.module.css";
+import fontStyles from "@/components/web-print/css/printers.fonts.module.css";
+import marginStyles from "@/components/web-print/css/printers.margins.module.css";
+import themeStyles from "@/components/web-print/css/printers.theme.module.css";
 import { Modal } from "@/components/common/Modal.tsx";
 
 export function FileDrop({
