+++ src/components/web-print/WebPrintPage.tsx
@@ -1,5 +1,5 @@
-import styles from "@/components/web-print/printers.module.css";
-import themeStyles from "@/components/web-print/printers.theme.module.css";
+import styles from "@/components/web-print/css/printers.module.css";
+import themeStyles from "@/components/web-print/css/printers.theme.module.css";
 import { DoubleScreenContainer } from "@/components/web-print/DoubleScreenContainer.tsx";
 import { ConfigurationScreen } from "@/components/web-print/ConfigurationScreen.tsx";
 import { ProcessingScreen } from "@/components/web-print/ProcessingScreen.tsx";
